WidgetVisibility
Implement Visibility for Elements.

Visibility should remove interactivity and visuals but keep the Element's space in the layout in the same way it would be if fully rendered.

This commit doesn't implement Visibility for children below a Hidden parent, although Layers provide part of this functionality since they control the opacity of their children.